Recieved data in REST service for Author data.User bean needs to be created for business logic

+++ model/LibraryEntities/src/main/java/org/training/example/models/Book.java
@@ -29,7 +29,7 @@ public class Book extends LibraryEntity {
     private String description;
 
     @JsonbProperty("isbn")
-    private int isbn;
+    private long isbn;
 
     @JsonbProperty("noOfPages")
     private int noOfPages;
@@ -64,11 +64,11 @@ public class Book extends LibraryEntity {
 	this.noOfPages = noOfPages;
     }
 
-    public int getIsbn() {
+    public long getIsbn() {
 	return isbn;
     }
 
-    public void setIsbn(int isbn) {
+    public void setIsbn(long isbn) {
 	this.isbn = isbn;
     }
